EF config changes

+++ src/SFA.DAS.Payments.Model.Core/Audit/EarningEventModel.cs
@@ -8,7 +8,7 @@ namespace SFA.DAS.Payments.Model.Core.Audit
         public long Id { get; set; }
         public ContractType ContractType { get; set; }
         public string AgreementId { get; set; }
-        public int LearningAimSequenceNumber { get; set; }
+        public long LearningAimSequenceNumber { get; set; }
         public List<EarningEventPeriodModel> Periods { get; set; }
         public List<EarningEventPriceEpisodeModel> PriceEpisodes { get; set; }
